% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{add_adjusted}
\alias{add_adjusted}
\title{Add expression data adjusted for pairs/surrogate variables}
\usage{
add_adjusted(eset, svobj = list(sv = NULL), numsv = 0)
}
\arguments{
\item{eset}{ExpressionSet}

\item{svobj}{surrogate variable object}

\item{numsv}{Number of surrogate variables to adjust for}
}
\value{
eset with \code{adjusted} element added
}
\description{
Add expression data adjusted for pairs/surrogate variables
}
