% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{add_contrasts}
\alias{add_contrasts}
\title{Select contrasts for each GSE.}
\usage{
add_contrasts(eset, gse_name, annot = "SYMBOL", prev_anal)
}
\arguments{
\item{eset}{single eSet passed to diff_expr.}

\item{gse_name}{Character, GSE name of eset.}

\item{annot}{Character, one of "PROBE" or "SYMBOL" for probe or gene level
analysis respectively. For duplicated gene symbols,the feature with
the highest interquartile range across selected samples will be kept.}

\item{prev_anal}{result of previous call to diff_expr}
}
\value{
list needed for \code{\link{diff_setup}}
}
\description{
Function is use by diff_expr to get sample selections for each contrast from
user. After selecting samples, duplicated features (probes or genes) are
removed.
}
\seealso{
\code{\link{diff_expr}}
}

