% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils_new.R
\name{add_es_new}
\alias{add_es_new}
\title{Add metaMA effectsize values to top table.}
\usage{
add_es_new(
  tt,
  ebfit,
  cols = c("dprime", "vardprime"),
  groups = c("test", "ctrl")
)
}
\arguments{
\item{cols}{Columns from \code{\link[metaMA]{effectsize}} result to add to
top table.}

\item{diff_exprs}{Result from call.}
}
\value{
diff_exprs with specified columns added to top_table.
}
\description{
Adds moderated unbiased standardised effect sizes (dprimes) to top table
from differential expression analysis.
}
\examples{
# location of previously saved analysis
data_dir <- system.file('extdata', 'IBD', package = 'drugseqr')

# load previous analysis for eset
anal <- readRDS(file.path(data_dir, 'diff_exprs_symbol.rds'))

# add dprime and vardprime to top tables
anal <- add_es(anal)
}
