% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_illum.R
\name{add_pvals}
\alias{add_pvals}
\title{Add detection p-values to Illumina expression set.}
\usage{
add_pvals(eset, pvals)
}
\arguments{
\item{eset}{Illumina expression set to add pvals slot to.}

\item{pvals}{in Detection slot obtained from \link{read.ilmn}}
}
\value{
expression slot with pvals slot containing detection p-values.
}
\description{
Adds detection p-vals to pvals slot of illumina expression set.
}
\examples{
\dontrun{

}
}

