% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_utils.R
\name{add_sources}
\alias{add_sources}
\title{Add sample source information for meta-analysis.}
\usage{
add_sources(diff_exprs, data_dir = getwd())
}
\arguments{
\item{diff_exprs}{Result of previous call to \code{diff_expr}.}
}
\value{
Same as \code{\link{diff_expr}} with added slots for each GSE in list:
  \item{sources}{Named vector specifying selected sample source for each contrast.
     Vector names identify the contrast.}
  \item{pairs}{List of character vectors specifying selected sample sources
     that should be analysed together for meta-analysis.}
}
\description{
Add sample source information for meta-analysis.
}
\examples{
blah <- c()
}
