% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-contrasts-ui.R
\name{bulkAnnotInput}
\alias{bulkAnnotInput}
\title{UI for Bulk Data annotation upload/download}
\usage{
bulkAnnotInput(id)
}
\description{
UI for Bulk Data annotation upload/download
}
\keyword{internal}
