% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-contrasts-ui.R
\name{bulkPageUI}
\alias{bulkPageUI}
\title{UI for Select Contrasts Interface}
\usage{
bulkPageUI(id)
}
\arguments{
\item{id}{The id string to be namespaced.}
}
\description{
UI for Select Contrasts Interface
}
