% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_affy.R
\name{cel_dates}
\alias{cel_dates}
\title{Extract run date from Affymetrix CEL file.}
\usage{
cel_dates(cel_paths)
}
\arguments{
\item{cel_paths}{Charactor vector, full paths to CEL files.}
}
\value{
Factor vector of CEL run dates.
}
\description{
Useful for defining sample batches for \code{\link{ComBat}}. No longer
used by crossmeta, which discovers nuissance variables using \code{\link{sva}}.
}
\seealso{
\code{\link{ComBat}}
}

