% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{ch2_subset}
\alias{ch2_subset}
\title{Subset for Paired Two-Channel ExpressionSet}
\usage{
ch2_subset(eset, prev_anal)
}
\arguments{
\item{eset}{Annotated ExpressionSet. Created by \code{load_raw}.}

\item{prev_anal}{One item (for eset) from previous result of \code{diff_expr}.}
}
\value{
ExpressionSet. If two-channel, paired and one channel not used will subset to used channel.
}
\description{
Two-channel esets use intraspotCorrelation and lmscFit so can't use duplicateCorrelation.
If not using one channel in contrasts (e.g. because all reference RNA) and have paired design,
better to treat as single channel so that can use duplicateCorrelation.
}
\keyword{internal}
