% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{clean_y}
\alias{clean_y}
\title{Adjustes expression data for surrogate variables.}
\usage{
clean_y(y, mod, svs)
}
\arguments{
\item{y}{expression data of eset.}

\item{mod}{full model matrix supplied to \link{sva}.}

\item{svs}{surrogate variables returned by \link{sva} (svobj$sv).}
}
\value{
expression data with effect of svs removed.
}
\description{
Factors out effect of surrogate variables discovered during analysis by
\link{sva}.
}
\seealso{
\code{\link{get_contrast_esets}}.
}

