% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{commonize}
\alias{commonize}
\title{Keep only common features for a list of esets.}
\usage{
commonize(esets, annot = "SYMBOL")
}
\arguments{
\item{esets}{list of annotated expression sets to commonize.}

\item{annot}{feature names to commonize by. Either "SYMBOL" or "PROBE".}
}
\value{
list of esets with where all features (probe or gene) are common.
}
\description{
Used prior to differential expression analysis (diff_expr) to remove
non-common features (either gene or probe name). The subsequent meta-analysis
uses common features only, so eliminating non-common features reduces the number
of comparisons made during differential expression analysis (increases power).
}
\examples{
\dontrun{

}
}
\seealso{
\link{load_affy}, \link{load_illum}, and \link{load_agil} to obtain
         list of annotated esets.
         \link{diff_expr} to run differential expression subsequent to
         \code{commonize}.
}

