% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{diff_anal}
\alias{diff_anal}
\title{Run limma analysis.}
\usage{
diff_anal(eset, contrasts, group_levels, mama_data, mod, modsv, svobj, gse_dir,
  gse_name, annot = "SYMBOL")
}
\arguments{
\item{eset}{eset, generated by \code{\link{add_contrasts}}.}

\item{contrasts}{Character vector, generated by \code{\link{add_contrasts}}.}

\item{group_levels}{Character vector, generated by \code{\link{add_contrasts}}.}

\item{mama_data}{List, generated by \code{\link{add_contrasts}}.}

\item{mod}{model matrix, generated by \code{\link{diff_setup}}.}

\item{modsv}{model matrix with surrogate variables,
generated by \code{\link{diff_setup}}.}

\item{svobj}{result from \code{\link{sva}} function called during
\code{\link{diff_setup}}.}

\item{gse_dir}{Character, path to directory with GSE data.}

\item{gse_name}{Character, name of GSE.}

\item{annot}{one of "SYMBOL" or "PROBE". Different save names used for each
so that can run analysis both at gene or probe level.}
}
\value{
List, final result of \code{diff_expr}. Used for subsequent meta-analysis.
}
\description{
Runs limma differential expression analysis on all contrasts selected by
\code{\link{add_contrasts}}. Analysis performed with and without surrogate
variables discovered by \code{\link{diff_setup}}. Also print MDS plot and saves
results to disc.
}
\seealso{
\code{\link{diff_expr}}.
}

