% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{diff_setup}
\alias{diff_setup}
\title{Generate model matrix with surrogate variables.}
\usage{
diff_setup(eset, group_levels)
}
\arguments{
\item{eset}{eset with "group" column in phenoData.}

\item{group_levels}{unique group names created by \code{\link{add_contrasts}}.}
}
\value{
list with model matrix(mod), model matrix with surrogate
        variables(modsv), and result of \code{\link{sva}} function.
}
\description{
Used by \code{\link{diff_expr}} to get model matrix with surrogate variabled
in order to run \code{\link{diff_anal}}.
}
\seealso{
\code{\link{add_contrasts}}, \code{\link{diff_expr}}.
}

