% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_edits.R
\name{EScombination}
\alias{EScombination}
\title{Edit of metaMA::\link{EScombination} for multiple contrasts per study.}
\usage{
EScombination(esets, classes, ebayes, BHth = 0.05)
}
\arguments{
\item{esets}{list of expression sets (one per contrast).}

\item{classes}{list of factor vectors indicating class labels for each contrast.}

\item{ebayes}{result of call to \code{get_ebayes}.}

\item{BHth}{cutoff for false discovery rate.}
}
\value{
List
        \item{Study1}{Vector of indices of differentially expressed genes in
         study 1. Similar names are given for the other individual studies.}
        \item{AllIndStudies}{Vector of indices of differentially expressed
        genes found by at least one of the individual studies.}
        \item{Meta}{Vector of indices of differentially expressed genes in
        the meta-analysis.}
        \item{TestStatistic}{Vector with test statistics for differential
        expression in the meta-analysis.}
}
\description{
Results from seperate differential expression can be supplied. This allows for
multiple contrasts to be analysed within a study.
}
\examples{
\dontrun{

}
}
\seealso{
\link{make_ma} to make MetaArray object from result of call to diff_expr.

         \link{GEDM} and \link{selectClass} to get esets and classes respectively.

         \link{get_ebayes} to extract ebayes info for each contrast.

         \link[metaMA]{pvalcombination} for original method.
}

