% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_path.R
\name{explore_paths}
\alias{explore_paths}
\title{Explore pathway meta analysis.}
\usage{
explore_paths(es, path_res, drug_info, type = c("both", "mimic", "reverse"))
}
\arguments{
\item{es}{Result of call to \code{\link{es_meta}}.}

\item{path_res}{Result of call to \code{\link{path_meta}}.}

\item{drug_info}{Character vector specifying which dataset to query
(either 'cmap' or 'l1000'). Can also provide a matrix of differential expression
values for drugs or drug combinations (rows are genes, columns are drugs).}

\item{type}{Desired direction of drug action on query signature. If \code{'both'},
drugs that mimic (positive cosines) and reverse (negative cosines) the query signature
are at the top and bottom of drug selection boxes respectively.}
}
\description{
Explore pathway meta analysis.
}
\examples{
blah <- c()
}
