% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter genes in RNA-seq ExpressionSet}
\usage{
filter_genes(eset)
}
\arguments{
\item{eset}{ExpressionSet with 'counts' assayDataElement and group column in
pData}
}
\value{
filtered \code{eset}
}
\description{
Uses \link[edgeR]{filterByExpr} to filter based on 'counts' assay or 'exprs'
assay if 'counts' isn't available (for ARCHS4 data).
}
\examples{

# example ExpressionSet
eset <- makeExampleCountsEset()
eset <- filter_genes(eset)

}
\seealso{
\link[edgeR]{filterByExpr}
}
