% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{fit_ebayes}
\alias{fit_ebayes}
\title{Fit ebayes model}
\usage{
fit_ebayes(lm_fit, contrasts, robust = TRUE)
}
\arguments{
\item{lm_fit}{Result of call to \link{run_limma}}

\item{contrasts}{Character vector of contrasts to fit.}

\item{robust}{logical, should the estimation of \code{df.prior} and \code{var.prior} be robustified against outlier sample variances?}
}
\value{
result of \link[limma]{eBayes}
}
\description{
Fit ebayes model
}
