% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils_new.R
\name{fit_ebayes_new}
\alias{fit_ebayes_new}
\title{Fit ebayes model}
\usage{
fit_ebayes_new(lm_fit, contrasts, robust = TRUE)
}
\arguments{
\item{lm_fit}{Result of call to \link{run_limma}}

\item{contrasts}{Character vector of contrasts to fit.}
}
\value{
result of \link[limma]{eBayes}
}
\description{
Fit ebayes model
}
