% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{fit_lm}
\alias{fit_lm}
\title{Run limma analysis.}
\usage{
fit_lm(eset, svobj = list(sv = NULL), numsv = 0, rna_seq = TRUE, quick = FALSE)
}
\arguments{
\item{eset}{Annotated eset created by \code{load_raw}. Replicate features and
non-selected samples removed by \code{iqr_replicates}.}

\item{svobj}{Surrogate variable analysis results. Returned from \link{run_sva}.}

\item{numsv}{Number of surrogate variables to model.}

\item{rna_seq}{Is this an RNA-seq \code{eset}? Default is \code{TRUE}.}

\item{quick}{\code{TRUE} to speed up single-cell pseudobulk with a faster but less 
accurate \code{\link[limma]{voomWithQualityWeights}}.}
}
\value{
list with slots:
  * \code{fit} Result of \link[limma]{lmFit}.
  * \code{mod} model matrix used for fit.
}
\description{
Runs limma differential expression analysis on all contrasts selected by
\code{add_contrast}. Analysis performed with and without surrogate
variables discovered by \code{diff_setup}. Also prints MDS plot and saves
results.
}
\keyword{internal}
