% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illum_headers.R
\name{fix_illum_headers}
\alias{fix_illum_headers}
\title{Attempts to fix Illumina raw data header}
\usage{
fix_illum_headers(elist_paths, eset = NULL)
}
\arguments{
\item{elist_paths}{Path to Illumina raw data files. Usually contain patterns:
non_normalized.txt, raw.txt, or _supplementary_.txt}

\item{eset}{ExpressionSet from \link{getGEO}.}
}
\value{
Character vector for \code{annotation} argument to \link[limma]{read.ilmn}. Fixed raw data files
  are saved with filename ending in _fixed.txt
}
\description{
Reads raw data files and tries to fix them up so that they can be loaded by \link[limma]{read.ilmn}.
}
