% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_path.R
\name{format_fdrs}
\alias{format_fdrs}
\title{Make FDR values readable}
\usage{
format_fdrs(fdrs)
}
\arguments{
\item{fdrs}{Numeric vector of false discovery rates}
}
\value{
Character vector of formated false discovery rates.
}
\description{
Used by \code{\link{explore_paths}} to make pathway FDR values readable
}
\examples{

fdrs <- c(0.011, 0.00111, 0.000001111, 0.10000001, 1, 1.00000)
format_fdrs(fdrs)
}
\keyword{internal}
