% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{get_biocpack}
\alias{get_biocpack}
\title{Downloads bioconductor package.}
\usage{
get_biocpack(biocpack_name)
}
\arguments{
\item{biocpack_name}{name of bioconductor package to download.}
}
\value{
NULL (downloads and loads requested package).
}
\description{
Used by symbol_annot to download annotation data packages from bioconductor.
}
\examples{
\dontrun{

}
}
\seealso{
\link{get_biocpack_name}, \link{symbol_annot}.
}

