% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{get_biocpack_name}
\alias{get_biocpack_name}
\title{Queries GEOmetadb sqlite file to obtain bioconductor annotation data package name.}
\usage{
get_biocpack_name(gpl_name)
}
\arguments{
\item{gpl_name}{platform name of GSE.}
}
\value{
name of bioconductor package.
}
\description{
Function queries GEOmetadb.sqlite in order to obtain bioconductor annotation
data package name for a given platform (GPL). If bioconductor package name
not available, user must look up manually on bioconductor website and then
enter.
}
\examples{
\dontrun{

}
}
\seealso{
\link{symbol_annot}.
}

