% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{get_ch2_mod}
\alias{get_ch2_mod}
\title{Get design matrix for two-channel array}
\usage{
get_ch2_mod(eset)
}
\arguments{
\item{eset}{ExpressionSet with \code{colnames} that end in '_red' and '_green'
indicating channel and \code{eset$group} indicating group membership.}
}
\value{
model matrix for use by \link[limma]{intraspotCorrelation} and \link[limma]{lmscFit}
}
\description{
Get design matrix for two-channel array
}
\keyword{internal}
