% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_utils.R
\name{get_ebayes}
\alias{get_ebayes}
\title{Get ebayes info for each contrast.}
\usage{
get_ebayes(diff_exprs, sva)
}
\arguments{
\item{diff_exprs}{result of previous call to \code{diff_expr} of \code{load_diff}.}

\item{sva}{Do you want ebayes info from model with surrogate variables?}
}
\value{
list of MArrayLM objects (one per contrast - contain ebayes info).
}
\description{
Helper utility in order to obtain ebayes info for each contrast.
}
\examples{
\dontrun{

}
}
\seealso{
\link{diff_expr}, or \link{load_diff} to obtain diff_exprs.

         \link{pvalcombination}, and \link{EScombination} for functions that
         require result of \code{get_ebayes}.
}

