% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_affy.R
\name{get_eset_names}
\alias{get_eset_names}
\title{Get eset names for load_affy.}
\usage{
get_eset_names(esets, gse_names)
}
\arguments{
\item{esets}{loaded by \code{load_affy}.}

\item{gse_names}{argument supplied to \code{load_affy}.}
}
\value{
Character vector, GSE names with GPL appended to distinguish multiple
        platforms per GSE.
}
\description{
Helper function to get around issue of a single GSE having multiple platforms
(and thus \link{getGEO} returns a list of esets). To distinguish these cases,
 the GPL platform is appended to the GSE name.
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{load_affy}}
}

