% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{get_homologene}
\alias{get_homologene}
\title{Get homologene data frame.}
\usage{
get_homologene(homologene_path)
}
\arguments{
\item{homologene_path}{}
}
\description{
Function loads and, if necessary, sets up homologene data frame.
}
\details{
Setup results in a dataframe with all entrez ids in one column ('entrez') and
the homologous human entrez ids in another column ('entrez_HS').
}

