% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_agil.R
\name{get_raw_agil}
\alias{get_raw_agil}
\title{Download and unpack Agilent microarray supplementary files from GEO.}
\usage{
get_raw_agil(gse_names, data_dir)
}
\arguments{
\item{gse_names}{Character vector of GSE names to download.}

\item{data_dir}{base data directory (a folder for each GSE will be created
here).}
}
\value{
NULL (for download/unpack only).
}
\description{
Downloads Agilent microarray supplementary files from GEO.
Files are stored in the supplied data directory under the GSE name.
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{get_raw_illum}}, \code{\link{get_raw_agil}}.
}

