% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{get_sva_mods}
\alias{get_sva_mods}
\title{Get model matrices for surrogate variable analysis}
\usage{
get_sva_mods(pdata)
}
\arguments{
\item{pdata}{\code{data.frame} of phenotype data with column \code{'group'}
and \code{'pair'} (optional).}
}
\value{
List with model matrix(mod) and null model matrix (mod0) used for \code{sva}.
}
\description{
Used by \code{add_adjusted} to create model matrix with surrogate variables.
}
