% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{get_top_table}
\alias{get_top_table}
\title{Get top table}
\usage{
get_top_table(
  lm_fit,
  groups = c("test", "ctrl"),
  with.es = TRUE,
  robust = FALSE
)
}
\arguments{
\item{lm_fit}{Result of \link{run_limma}}

\item{groups}{Test and Control group as strings.}
}
\value{
result of \link[limma]{toptable}
}
\description{
Get top table
}
