% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_utils.R
\name{get_tts}
\alias{get_tts}
\title{Get topTables for each contrast.}
\usage{
get_tts(diff_exprs)
}
\arguments{
\item{diff_exprs}{result of call to \code{diff_expr} or \code{load_diff}.}
}
\value{
list of topTables (one for each contrast in diff_exprs).
}
\description{
Helper function for \code{merge_ranks} meta-analysis method.
}
\examples{
\dontrun{

}
}
\seealso{
\link{merge_ranks}.
}

