% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs.names}
\alias{gs.names}
\title{Map between KEGG pathway numbers and names.}
\format{An object of class \code{character} of length 310.}
\usage{
data(gs.names)
}
\value{
A named character vector of human KEGG pathway names. Names of vector are KEGG pathway numbers.
}
\description{
Used to map human KEGG pathway numbers to names. Updated Feb 2017.
}
\keyword{datasets}
