% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gslist}
\alias{gslist}
\title{KEGG human pathway genes.}
\format{An object of class \code{list} of length 312.}
\usage{
data(gslist)
}
\value{
A named list with entrez ids of genes for human KEGG pathways. List names are KEGG pathway numbers.
}
\description{
Genes for human KEGG pathways. Updated Feb 2017.
}
\keyword{datasets}
