% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illum_headers.R
\name{ilmn.nnum}
\alias{ilmn.nnum}
\title{Count numeric columns in raw Illumina data files}
\usage{
ilmn.nnum(elist_paths)
}
\arguments{
\item{elist_paths}{Paths to raw illumina data files}
}
\value{
Number of numeric columns in \code{elist_paths} excluding probe ID columns.
}
\description{
Excludes probe ID cols
}
