% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{inputs}
\alias{inputs}
\title{Query user to provide description.}
\usage{
inputs(msg = "", box1 = "eg. AL.obob", def1 = "AL",
  box2 = "eg. CR.obob", def2 = "CR", two = F)
}
\arguments{
\item{msg}{Message to display in title bar.}

\item{box1}{Description beside box1.}

\item{def1}{Default value for box1.}

\item{box2}{Description beside box2.}

\item{def2}{Default value for box2.}

\item{two}{Do you want two input boxes (one if FALSE)?}
}
\value{
Character vector with inputs typed into box1 and/or box2.
}
\description{
Uses tcltk to request input from user for group or tissue names.
}
\seealso{
\link{add_contrasts}
}

