% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{iqr_duplicates}
\alias{iqr_duplicates}
\title{Removes features with duplicated annotation.}
\usage{
iqr_duplicates(eset, annot = "SYMBOL")
}
\arguments{
\item{eset}{expression set with fData column "SYMBOL" or "PROBE".}

\item{annot}{Character, one of "PROBE" or "SYMBOL" for probe or gene level
analysis respectively. For duplicated gene symbols,the feature with
the highest interquartile range across selected samples will be kept.}
}
\value{
expression set with unique features at probe or gene level.
}
\description{
For rows with duplicated annot, highested IQR retained.
}

