% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_edits.R
\name{join.DEG}
\alias{join.DEG}
\title{Fixes bug in MAMA join.DEG.}
\usage{
join.DEG(..., genenames = NULL, type = NULL, cutoff)
}
\arguments{
\item{...}{Outputs from different function for methods of meta-analysis of
microarray}

\item{genenames}{a character vector - names of all genes (or probe ID)
included in meta-analysis. It can be NULL if the wrapper functions were
used for the analysis.}

\item{type}{a numeric vector idicating from which function the output is,
kth element in type corresponds to kth element of .... It is not
needed when wrapper functions where used.}

\item{cutoff}{a numeric value - a cutoff level for p-value to select
significant genes}
}
\value{
A list in which each slot refers to one meta-analytical method and
        contains names of differentially expressed genes found by the method.
}
\description{
Bug was present under condition \code{if (type[i] == 3)}. Column was selected
by number, but is not in general constant. Fix refers to column by name
("FDR").
}
\examples{
\dontrun{

}
}
\seealso{
\link[MAMA]{join.DEG} for original method.
}

