% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_affy.R
\name{load_affy}
\alias{load_affy}
\title{Load and pre-process raw Affymetrix CEL files for multiple GSEs.}
\usage{
load_affy(gse_names, data_dir)
}
\arguments{
\item{gse_names}{Character vector of GSE names to load and pre-process.}

\item{data_dir}{Character, base data directory (contains a folder with raw
data for each GSE to be loaded).}
}
\value{
list of processed esets (one for each unique GSE/GPL platform).
}
\description{
Load raw CEL files previously downloaded with \code{get_raw_affy}. Data is
RMA normalized, CEL scan dates are extracted and added to phenoData slots,
and gene SYMBOLs are added to featureData slots.
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{get_raw_affy}} to obtain raw Affymetrix CEL files for
         multiple GSEs.
}

