% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_affy.R
\name{load_affy_plat}
\alias{load_affy_plat}
\title{Helper utility for load_affy.}
\usage{
load_affy_plat(eset, gse_dir)
}
\arguments{
\item{eset}{GSEMatrix obtained by load_affy call to getGEO.}

\item{gse_dir}{directory containing raw data for GSE.}
}
\value{
eset with scan_date in pData slot and SYMBOL in fData slot.
}
\description{
Used by load_affy to load an eset for each GPL platform in a GSE.
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{load_affy}}
}

