% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_agil.R
\name{load_agil}
\alias{load_agil}
\title{Load and pre-process raw Agilent files for multiple GSEs.}
\usage{
load_agil(gse_names, data_dir)
}
\arguments{
\item{gse_names}{Character vector of GSE names to load and pre-process.}

\item{data_dir}{Character, base data directory (contains a folder with raw
data for each GSE to be loaded).}
}
\value{
list of processed esets.
}
\description{
Load raw txt files previously downloaded with \code{get_raw_agil}. Data is
neqc normalized and gene SYMBOLs are added to featureData slots.
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{get_raw_agil}} to obtain raw Agilent txt files for
         multiple GSEs.
}

