% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{load_diff}
\alias{load_diff}
\title{Load previous differential expression analyses.}
\usage{
load_diff(gse_names, data_dir = getwd(), annot = "SYMBOL")
}
\arguments{
\item{gse_names}{Character vector specifying GSE names to be loaded.}

\item{data_dir}{String specifying directory of GSE folders.}

\item{annot}{Level of previous analysis (e.g. "SYMBOL" or "PROBE").}
}
\value{
Result of previous call to \code{\link{diff_expr}}.
}
\description{
Loads previous differential expression analyses.
}
\examples{
library(lydata)

data_dir <- system.file("extdata", package = "lydata")
gse_names <- c("GSE9601", "GSE34817")
prev <- load_diff(gse_names, data_dir)
}
