% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_illum.R
\name{load_illum}
\alias{load_illum}
\title{Load and pre-process raw Illum files for multiple GSEs.}
\usage{
load_illum(gse_names, data_dir)
}
\arguments{
\item{gse_names}{Character vector of GSE names to load and pre-process.}

\item{data_dir}{Character, base data directory (contains a folder with raw
data for each GSE to be loaded).}
}
\value{
list of processed esets.
}
\description{
Load raw txt files previously downloaded with \code{get_raw_illum} and checked
for format (see \code{\link{open_raw_illum}}). Data is normalized using neqc,
detection p-values are added to pvals slot and gene SYMBOLs are added to
featureData slot.
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{get_raw_illum}} to obtain raw Illumina files and
         \code{\link{open_raw_illum}} to ensure their correctness.
}

