% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_illum.R
\name{load_illum_plat}
\alias{load_illum_plat}
\title{Illumina loader utility for load_plat.}
\usage{
load_illum_plat(eset, gse_name, gse_dir, ensql)
}
\arguments{
\item{eset}{Expression set obtained by \code{getGEO}.}

\item{gse_name}{String specifying GSE name.}

\item{gse_dir}{String specifying path to GSE folder.}
}
\value{
Annotated eset.
}
\description{
Used by load_plat to load an eset.
}
\seealso{
\code{\link{load_plat}}.
}
\keyword{internal}
