% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{load_path}
\alias{load_path}
\title{Load previous pathway analyses.}
\usage{
load_path(gse_names, data_dir = getwd())
}
\arguments{
\item{gse_names}{Character vector of GSE names.}

\item{data_dir}{String specifying directory for GSE folders.}
}
\value{
Result of previous call to \code{diff_path}.
}
\description{
Load previous pathway analyses.
}
\examples{

library(lydata)

# location of data
data_dir <- system.file("extdata", package = "lydata")

# gather GSE names
gse_names  <- c("GSE9601", "GSE15069", "GSE50841", "GSE34817", "GSE29689")

# load previous pathway analyses
path_anals <- load_path(gse_names, data_dir)
}
