% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{load_plat}
\alias{load_plat}
\title{Load and pre-process raw Affymetrix, Illumina, and Agilent microarrays.}
\usage{
load_plat(gse_name, data_dir, gpl_dir, ensql)
}
\arguments{
\item{gse_name}{GSE names.}

\item{data_dir}{String specifying directory with GSE folder.}

\item{gpl_dir}{String specifying parent directory to search for previously downloaded GPL.soft files.}

\item{ensql}{For development. Path to sqlite file with ENTREZID and SYMBOL columns created in data-raw/entrezdt.}
}
\value{
List of annotated esets, one for each platform in \code{gse_name}.
}
\description{
Load raw files previously downloaded with \code{get_raw}. Used by \code{load_raw}.
}
\details{
Data is normalized, SYMBOL and PROBE annotation are added to fData slot.
}
\seealso{
\code{\link{get_raw}} to obtain raw data.
}
\keyword{internal}
