% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_utils.R
\name{make_ma}
\alias{make_ma}
\title{Make MetaArray object from results of differential expression.}
\usage{
make_ma(diff_exprs, sva)
}
\arguments{
\item{diff_exprs}{result of call to \code{diff_expr} or \code{load_diff}.}

\item{sva}{Use esets with effect of surrogate variables removed? If FALSE,
none-adjusted esets will be used.}
}
\value{
MetaArray object.
}
\description{
Function is used to construct a MetaArray object from results of call to
\code{diff_expr} or \code{load_diff}. MetaArray object allows for use of
all meta-analysis methods present in MAMA package.
}
\seealso{
\link[MAMA]{MetaArray-class} for possible meta-analysis methods.

         \link{diff_expr} or \link{load_diff} to obtain \code{diff_expr}.
}

