% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes.R
\name{makeExampleCountsEset}
\alias{makeExampleCountsEset}
\title{Make example ExpressionSet}
\usage{
makeExampleCountsEset(
  n = 1000,
  m = 12,
  betaSD = 0,
  interceptMean = 4,
  interceptSD = 2,
  dispMeanRel = function(x) 4/x + 0.1,
  sizeFactors = rep(1, m)
)
}
\arguments{
\item{n}{number of rows}

\item{m}{number of columns}

\item{betaSD}{the standard deviation for non-intercept betas, i.e. beta ~ N(0,betaSD)}

\item{interceptMean}{the mean of the intercept betas (log2 scale)}

\item{interceptSD}{the standard deviation of the intercept betas (log2 scale)}

\item{dispMeanRel}{a function specifying the relationship of the dispersions on
\code{2^trueIntercept}}

\item{sizeFactors}{multiplicative factors for each sample}
}
\description{
adapted from DESeq2::makeExampleDESeqDataSet
}
\examples{
eset <- makeExampleCountsEset()

}
