% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{match_prev_eset}
\alias{match_prev_eset}
\title{Reuse contrast selections from previous analysis.}
\usage{
match_prev_eset(eset, prev_anal)
}
\arguments{
\item{eset}{Annotated eset. Created by \code{load_raw}.}

\item{prev_anal}{One item (for eset) from previous result of \code{diff_expr}.
If present, previous selections and names will be reused.}
}
\value{
Expression set with samples and pData as in prev_anal.
}
\description{
Transfers user-supplied selections from previous call of diff_expr.
}
\seealso{
\code{\link{diff_expr}}
}
\keyword{internal}
