% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_utils.R
\name{merge_ranks}
\alias{merge_ranks}
\title{Employ RankMerging meta-analysis from GeneExpressionSignature package.}
\usage{
merge_ranks(diff_exprs, n = NULL)
}
\arguments{
\item{diff_exprs}{result of call to \code{diff_expr} or \code{load_diff}.}

\item{n}{number of top ranked (up-regulated) and bottom ranked (down-regulated)
genes to return.}
}
\value{
Character vector of gene identifiers sorted by rank.
}
\description{
Wrapper for RankMerging function in GeneExpressionSignature.
}
\examples{
\dontrun{

}
}
\seealso{
\link{RankMerging}, \link{diff_expr}, \link{load_diff}.
}

