% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_illum.R
\name{open_raw_illum}
\alias{open_raw_illum}
\title{Opens non-normalized Illumina txt and xls files.}
\usage{
open_raw_illum(gse_names, data_dir)
}
\arguments{
\item{gse_names}{Character vector of GSE names to open raw files for.}

\item{data_dir}{base data directory (contains folder for each GSE).}
}
\value{
Character vector of successfully formatted GSEs (only load these).
}
\description{
Helper utility to open non-normalized Illumina files. User must check and
possibly edit the files. To help, I recommend downloading Sublime Text 2
(a text editor with regular expression capabilities).
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{load_illum}}, \code{\link{get_raw_illum}}.
}

