% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{path_meta}
\alias{path_meta}
\title{Fisher's p-value combination meta analysis of PADOG pathway analyses.}
\usage{
path_meta(path_anals, ncores = parallel::detectCores(), nperm = ncores *
  10000, by_source = FALSE)
}
\arguments{
\item{path_anals}{Result of previous call to \code{diff_path} or \code{load_path}.}

\item{ncores}{Number of cores to use. Default is all available.}

\item{nperm}{Number of permutation to perform to calculate p-values.}
}
\value{
Matrix with PADOG p-values for each contrast and permutation p- and fdr-values for meta analysis.
}
\description{
\code{\link[metap]{sumlog}} is used to perform meta-analysis of \code{\link[PADOG]{padog}} p-values.
Permutation p-values are determined by shuffling pathway names associated with PADOG p-values prior
to meta-analysis. Permutation p-values are adjusted using the Benjamini & Hochberg method to  obtain
false discovery rates.
}
\examples{

library(lydata)

# location of data
data_dir <- system.file("extdata", package = "lydata")

# gather GSE names
gse_names  <- c("GSE9601", "GSE15069", "GSE50841", "GSE34817", "GSE29689")

# load previous pathway analyses
path_anals <- load_path(gse_names, data_dir)

# perform pathway meta analysis
path_res <- path_meta(path_anals, ncores = 1, nperm = 100)

}
