% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_agil.R
\name{phenoData.ch2}
\alias{phenoData.ch2}
\title{Construct AnnotatedDataFrame from Two-Channel ExpressionSet}
\usage{
phenoData.ch2(eset)
}
\arguments{
\item{eset}{ExpressionSet with \code{pData} for two-channel Agilent array.}
}
\value{
AnnotatedDataFrame with twice as many rows as \code{eset}, one for each channel of each array in order all red then all green.
}
\description{
Construct AnnotatedDataFrame from Two-Channel ExpressionSet
}
