% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{plot_path}
\alias{plot_path}
\title{Plot meta analysis of pathway.}
\usage{
plot_path(path, es, drug_info = NULL, drugs = NULL, xlim = NULL)
}
\arguments{
\item{path}{Character vector giving pathway name. Valid values are row.names of path_meta result.}

\item{es}{Result of call to \code{es_meta}.}

\item{drug_info}{Matrix of expression values that include columns named for each drug.}

\item{drugs}{Character vector giving drug names to plot alongside meta analysis result.}

\item{xlim}{Numeric vector of length 2 specifying lower and upper bounds for dprime values.}
}
\value{
plot
}
\description{
For all transcripts in a given pathway, dprime values are plotted as grey circles for each contrast. Black
error bars are also plotted to indicate +/- one standard deviation from the overall mean effect sizes. The
parameters \code{drugs} and \code{drug_info} are used to plot
}
\examples{
library(ccmap)
library(ccdata)
library(lydata)

data_dir <- system.file("extdata", package = "lydata")
data(cmap_es)

# gather GSE names
gse_names  <- c("GSE9601", "GSE15069", "GSE50841", "GSE34817", "GSE29689")

# load previous differential expression analysis
anals <- load_diff(gse_names, data_dir)

# run meta-analysis
es <- es_meta(anals)

# highest ranking pathway from path_meta
path <- 'Amino sugar and nucleotide sugar metabolism'

# plot meta-analysis of pathway
plot_path(path, es, cmap_es)

# get meta-analysis effect size values
dprimes <- get_dprimes(es)

# drug with most similar transcriptional profile
topd <- names(query_drugs(dprimes$all$meta, cmap_es))[1]

# drug with most similar transcriptional profile for top pathway
topd_path <- names(query_drugs(dprimes$all$meta, cmap_es, path=path))[1]

# plot meta-analysis of pathway alongside top overall drug and top drug for pathway
plot_path(path, es, cmap_es, c(topd, topd_path))
}
