% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illum_headers.R
\name{prefix_illum_headers}
\alias{prefix_illum_headers}
\title{Run prefix on Illumina raw data files}
\usage{
prefix_illum_headers(elist_paths)
}
\arguments{
\item{elist_paths}{Paths to raw Illumina data files}
}
\value{
Paths to fixed versions of \code{elist_paths}
}
\description{
Run prefix on Illumina raw data files
}
