% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_illum.R
\name{query_ref}
\alias{query_ref}
\title{Get correlation between query and reference signatures.}
\usage{
query_ref(query, ref, sorted = TRUE, ngenes = 200)
}
\arguments{
\item{query}{Named numeric vector of differentual expression values for
query genes. Usually 'meta' slot of \code{get_dprimes} result.}

\item{ref}{A matrix of differential expression
to query against (rows are genes, columns are samples).}

\item{sorted}{Would you like the results sorted by decreasing similarity?
Default is TRUE.}

\item{ngenes}{The number of top differentially-regulated (up and down) query genes to use.}
}
\value{
Vector of pearson correlations between query and reference signatures.
}
\description{
Determines the pearson correlation between the query and each reference signature.
}
\keyword{internal}
