% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{run_limma}
\alias{run_limma}
\title{Linear model fitting of eset with limma.}
\usage{
run_limma(
  eset,
  annot = "SYMBOL",
  svobj = list(sv = NULL),
  numsv = 0,
  filter = TRUE
)
}
\arguments{
\item{eset}{Annotated eset created by \code{load_raw}.}

\item{annot}{String, column name in fData. For duplicated
values in this column, the row with the highest interquartile range
across selected samples will be kept. Appropriate values are \code{"SYMBOL"} (default - for gene level analysis)
or \code{"ENTREZID_HS"} (for probe level analysis).}

\item{svobj}{Surrogate variable analysis results. Returned from \link{run_sva}.}

\item{numsv}{Number of surrogate variables to model.}

\item{filter}{For RNA-seq. Should genes with low counts be filtered? dseqr shiny app performs this step
separately. Should be \code{TRUE} (default) if used outside of dseqr shiny app.}
}
\value{
List with:
  \item{fit}{result of \code{\link[limma]{lmFit}}.}
  \item{mod}{\code{model.matrix} used for \code{fit}}
}
\description{
After selecting control and test samples for a contrast, surrogate variable
analysis (\code{\link[sva]{sva}}) and linear model fitting with \link[limma]{lmFit} is performed.
}
\details{
If analyses need to be repeated, previous results can be reloaded with \code{\link[base]{readRDS}}
and supplied to the \code{prev_anal} parameter. In this case, previous selections will be reused.
}
