% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{run_limma_setup}
\alias{run_limma_setup}
\title{Setup ExpressionSet for running limma analysis}
\usage{
run_limma_setup(eset, prev)
}
\arguments{
\item{eset}{ExpressionSet}

\item{prev}{previous result of call to diff_expr}
}
\value{
\code{eset} ready for \code{run_limma}
}
\description{
Setup ExpressionSet for running limma analysis
}
