% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_select_contrasts.R
\name{run_select_contrasts}
\alias{run_select_contrasts}
\title{Shiny gadget to upload groups and select contrasts}
\usage{
run_select_contrasts(
  eset,
  app_dir = system.file("select_contrasts", package = "crossmeta", mustWork = TRUE),
  port = 3838
)
}
\arguments{
\item{eset}{ExpressionSet}

\item{app_dir}{Directory to shiny app. For local development use 'inst/select_contrasts'. Default is
in 'select_contrasts' sub directory of crossmeta package.}
}
\value{
result of \link{setup_prev}. Used to specify sample groups and contrasts for differential expression analysis.
}
\description{
Shiny gadget to upload groups and select contrasts
}
