% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils_new.R
\name{run_sva_new}
\alias{run_sva_new}
\title{Run surrogate variable analysis}
\usage{
run_sva_new(mods, eset, svanal = TRUE)
}
\arguments{
\item{mods}{result of \code{get_sva_mods}}

\item{eset}{Expression set.}

\item{rna_seq}{Boolean is it RNA-seq?}
}
\description{
Run surrogate variable analysis
}
