% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{symbol_annot}
\alias{symbol_annot}
\title{Add hgnc symbol to expression set.}
\usage{
symbol_annot(eset, gse_name = "", ensql = NULL)
}
\arguments{
\item{eset}{Expression set to annotate.}

\item{gse_name}{GSE name for eset.}
}
\value{
Expression set with hgnc symbols ("SYMBOL") and row names ("PROBE")
   added to fData slot.
}
\description{
Function first maps entrez gene ids to homologous human entrez gene ids and
then to hgnc symbols.
}
\details{
Initial entrez gene ids are obtained from bioconductor annotation data
packages or from feature data of supplied expression set. Homologous human
entrez ids are obtained from homologene and then mapped to hgnc symbols
using org.Hs.eg.db. Expression set is expanded if 1:many mappings occur.
}
\examples{
library(lydata)

# location of raw data
data_dir <- system.file("extdata", package = "lydata")

# load eset
eset <- load_raw("GSE9601", data_dir)[[1]]

# annotate eset (need if load_raw failed to annotate)
eset <- symbol_annot(eset)
}
\seealso{
\code{\link{load_raw}}.
}
