% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_agil.R
\name{to_eset}
\alias{to_eset}
\title{Convert limma object to ExpressionSet}
\usage{
to_eset(object, eset)
}
\arguments{
\item{object}{an EList of MAList object containing expression data.}

\item{eset}{ExpressionSet from \link{getGEO}. Used for annotation.}
}
\value{
ExpressionSet using expression data from \code{object} and annotation from \code{eset}.
}
\description{
Convert limma object to ExpressionSet
}
