% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_agil.R
\name{to_ma}
\alias{to_ma}
\title{Covert expression values to MAList}
\usage{
to_ma(y)
}
\arguments{
\item{y}{Expression values from two-channel agilent array in order all red then all green.}
}
\value{
MAList
}
\description{
Covert expression values to MAList
}
\examples{

A <- matrix(rnorm(100), ncol = 5)
M <- matrix(rnorm(100), ncol = 5)
MA <- new('MAList', list(M=M, A=A))
colnames(MA) <- letters[1:5]

y <- exprs.MA(MA)
MA2 <- crossmeta:::to_ma(y)
all.equal(MA, MA2)

}
\keyword{internal}
