% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{which_max_iqr}
\alias{which_max_iqr}
\title{Get row indices of maximum IQR within annotation groups}
\usage{
which_max_iqr(eset, groub_by, x = exprs(eset))
}
\arguments{
\item{eset}{\code{ExpressionSet}}

\item{groub_by}{Column in \code{fData(eset)} to group by}

\item{x}{\code{matrix} of expression values to use for IQR}
}
\value{
Integer vector of row numbers representing rows with the maximum IQR after grouping by \code{group_by}
}
\description{
Groups by \code{group_by} and determines row with maximum IQR.
}
