% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_illum.R
\name{xls_to_txt}
\alias{xls_to_txt}
\title{Covert .xls files to .txt}
\usage{
xls_to_txt(xls_paths)
}
\arguments{
\item{xls_paths}{Paths to .xls files}
}
\description{
For converting Illumina _Supplementary_.*.xls files to .txt for load_illum_plat.
}
\keyword{internal}
