% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{add_vsd}
\alias{add_vsd}
\title{Add VST normalized assay data element to expression set}
\usage{
add_vsd(eset, rna_seq = TRUE, pbulk = FALSE, vsd_path = NULL)
}
\arguments{
\item{eset}{ExpressionSet with group column in \code{pData(eset)}}

\item{rna_seq}{Is this an RNA-seq \code{eset}? Default is \code{TRUE}.}

\item{pbulk}{Is this an pseudobulk single-cell \code{eset}? Default is \code{FALSE}.
Used by package drugseqr.}

\item{vsd_path}{Path to save result to. Allows skipping running transform
on each load.}
}
\value{
\code{eset} with \code{'vsd'} \code{assayDataElement} added.
}
\description{
For microarray datasets duplicates exprs slot into vsd slot.
}
