% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{contribute}
\alias{contribute}
\title{Contribute results of meta-analysis to public database.}
\usage{
contribute(diff_exprs, subject)
}
\arguments{
\item{diff_exprs}{Result of call to \code{diff_expr}.}

\item{subject}{String identifying meta-analysis subject (e.g. "rapamycin" or
"prostate_cancer").}
}
\description{
Contributed results will be used to build a freely searchable database of
gene expression meta-analyses.
}
\details{
Performs meta-analysis using \code{es_meta} for . Sends overall mean effect size
values and minimal information needed to reproduce meta-analysis.
}
\examples{
library(lydata)

# location of data
data_dir <- system.file("extdata", package = "lydata")

# gather GSE names
gse_names  <- c("GSE9601", "GSE15069", "GSE50841", "GSE34817", "GSE29689")

# load differential expression analyses
anals <- load_diff(gse_names, data_dir)

# contribute results of meta-analysis
# contribute(anals, subject = "LY294002")
}

