% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{diff_expr}
\alias{diff_expr}
\title{Differential expression of list of eSets.}
\usage{
diff_expr(esets, data_dir, annot = "SYMBOL", prev_anals = list(NULL))
}
\arguments{
\item{data_dir}{Character, data directory containing GSE folders.}

\item{annot}{Character, one of "PROBE" or "SYMBOL" for probe or gene level
analysis respectively. For duplicated genes symbols,the feature with
the highest interquartile range across selected samples will be kept.}

\item{prev_anals}{previous output to re-use previous contrast selections.}

\item{eSet}{list from load_affy, load_illum, or load_agil.}
}
\value{
list needed for subsequent meta-analysis.
}
\description{
User selects contrasts, then surrogate variable analysis (sva) and
differential expression (limma) is run.
}

