% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{diff_expr}
\alias{diff_expr}
\title{Differential expression of esets.}
\usage{
diff_expr(esets, data_dir = getwd(), annot = "SYMBOL",
  prev_anals = list(NULL))
}
\arguments{
\item{esets}{List of annotated esets. Created by \code{load_raw}.}

\item{data_dir}{String specifying directory of GSE folders.}

\item{annot}{String, column name in fData common to all esets. For duplicated
values in this column, the row with the highest interquartile range
across selected samples will be kept. If meta-analysis will follow, appropriate
values are "SYMBOL" (default - for gene level analysis) or, if all esets are
from the same platform, "PROBE" (for probe level analysis).}

\item{prev_anals}{Previous result of \code{diff_expr}. If Present, previous
selections and names will be reused.}
}
\value{
List of lists (one per GSE), each containing:
  \item{eset}{Expression set without expression or feature data.
     \code{Treatment} ('ctl' or 'test') and \code{group} columns have been
     added to the \code{pData} slot. Only selected samples kept.}
  \item{top_tables}{List with results of \code{\link{topTable}} call (one per
     contrast). These results account for the effects of nuissance variables
     discovered by surrogate variable analysis.}
  \item{ebayes_sv}{Results of call to \code{\link{eBayes}} with surrogate
     variables included in the model matrix.}
}
\description{
User selects contrasts, then surrogate variable analysis (sva) and
differential expression analysis (limma) is performed.
}
\details{
For each GSE, analysis results are saved in the corresponding GSE
folder (in \code{data_dir}) that was created by \code{get_raw}. If analysis
needs to be repeated, previous results can be reloaded with \code{load_diff}
and supplied to the \code{prev_anals} parameter. In this case, previous
selections/names will be reused.
}
\examples{
library(lydata)

# location of raw data
data_dir <- system.file("extdata", package = "lydata")

# gather GSE names
gse_names  <- c("GSE9601", "GSE15069", "GSE50841", "GSE34817", "GSE29689")

# load first eset
esets <- load_raw(gse_names[1], data_dir)

# run analysis
# anals <- diff_expr(esets, data_dir)

# re-run analysis on first eset
prev <- load_diff(gse_names[1], data_dir)
anals <- diff_expr(esets[1], data_dir, prev_anals = prev)
}
\seealso{
\code{\link{get_raw}}, \code{\link{load_raw}}, and
  \code{\link{load_diff}}.
}

