% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_utils.R
\name{diff_path}
\alias{diff_path}
\title{Differential expression of KEGG pathways.}
\usage{
diff_path(esets, prev_anals, data_dir = getwd())
}
\arguments{
\item{esets}{List of annotated esets. Created by \code{load_raw}.}

\item{prev_anals}{Previous result of \code{diff_expr}.}

\item{data_dir}{String specifying directory of GSE folders. Defaul is working directory.}
}
\value{
List of data.frames containing PADOG pathway analysis results for each contrast.
}
\description{
Performs PADOG pathway analysis using KEGG database (downloaded Feb 2017).
}
\details{
Only contrasts with more than 2 samples per group can be analysed by PADOG.
}
\examples{

library(lydata)

# location of data
data_dir <- system.file("extdata", package = "lydata")

# gather GSE names
gse_names  <- c("GSE9601", "GSE15069", "GSE50841", "GSE34817", "GSE29689")

# load esets
esets <- load_raw(gse_names, data_dir)

# load previous differential expression analysis
anals <- load_diff(gse_names, data_dir)

# perform pathway analysis for each contrast
# path_anals <- diff_path(esets, anals, data_dir)
}
\seealso{
\code{\link[PADOG]{padog}} for details on PADOG pathway analysis.
}
