% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{es_meta}
\alias{es_meta}
\title{Effect size combination meta anaylsis.}
\usage{
es_meta(diff_exprs, cutoff = 0.3)
}
\arguments{
\item{diff_exprs}{Result of previous call to \code{diff_expr}.}

\item{cutoff}{Minimum fraction of contrasts that must have measured each gene.
Between 0 and 1.}
}
\value{
A data.frame with columns:
   \item{dprime}{Unbiased effect sizes(one column per contrast).}
   \item{vardprime}{Variances of unbiased effect sizes (one column per contrast).}
   \item{mu}{Overall mean effect sizes.}
   \item{var}{Variances of overall mean effect sizes.}
   \item{z}{Overall z score = \code{mu / sqrt(var)}.}
   \item{fdr}{False discovery rates calculated by \code{fdrtool}.}
}
\description{
Method builds on GeneMeta implementation by allowing for genes that were not
measured in all studies.
}
\details{
Method uses moderated effect sizes calculated by metaMA and determines fdr
using fdrtool.
}
\seealso{
\code{\link[GeneMeta]}, \code{\link[metaMA]{effectsize}},
   \code{\link{fdrtool}}.
}

