% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_utils.R
\name{es_meta}
\alias{es_meta}
\title{Effect size combination meta anaylsis.}
\usage{
es_meta(diff_exprs, cutoff = 0.3, by_source = FALSE)
}
\arguments{
\item{diff_exprs}{Result of previous call to \code{diff_expr}.}

\item{cutoff}{Minimum fraction of contrasts that must have measured each gene.
Between 0 and 1.}
}
\value{
A list with two named data.frames. The first ('filt') has all the
   columns below for genes present in cutoff or more fraction of contrasts.
   The second ('raw') has only \code{dprime} and \code{vardprime} columns but
   for all genes (NAs for genes not measured by a given contrast).
   \item{dprime}{Unbiased effect sizes (one column per contrast).}
   \item{vardprime}{Variances of unbiased effect sizes (one column per contrast).}
   \item{mu}{Overall mean effect sizes.}
   \item{var}{Variances of overall mean effect sizes.}
   \item{z}{Overall z score = \code{mu / sqrt(var)}.}
   \item{fdr}{False discovery rates calculated by \code{fdrtool}.}
}
\description{
Builds on GeneMeta implementation by allowing for genes that were not
measured in all studies.
}
\details{
In addition to allowing for genes that were not measured in all studies, this
method uses moderated unbiased effect sizes calculated by \code{metaMA} and
determines false discovery rates using \code{fdrtool}.
}
\examples{

library(lydata)

# location of data
data_dir <- system.file("extdata", package = "lydata")

# gather GSE names
gse_names  <- c("GSE9601", "GSE15069", "GSE50841", "GSE34817", "GSE29689")

# load previous analysis
anals <- load_diff(gse_names, data_dir)

# perform meta-analysis
es <- es_meta(anals)
}
\seealso{
\code{\link[GeneMeta]{zScores}}, \code{\link[metaMA]{effectsize}},
   \code{\link{fdrtool}}.
}
