% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{fit_ebayes}
\alias{fit_ebayes}
\title{Perform eBayes analysis from limma.}
\usage{
fit_ebayes(eset, contrasts, mod)
}
\arguments{
\item{eset}{eset, generated by \code{\link{add_contrasts}}.}

\item{contrasts}{Character vector, generated by \code{\link{add_contrasts}}.}

\item{mod}{model matrix (with or without surrogate variables modeled),
generated by \code{\link{diff_setup}}.}
}
\value{
result from called to limma \code{\link{eBayes}}.
}
\description{
Generates contrast matrix then runs eBayes analysis from limma.
}

