% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{get_contrast_esets}
\alias{get_contrast_esets}
\title{Generate eset for each contrast.}
\usage{
get_contrast_esets(sample_names, eset, data = exprs(eset))
}
\arguments{
\item{sample_names}{Character vector, sample names to include in eset.}

\item{eset}{eset to subset.}

\item{data}{expression data for eset (sva-adjusted or not).}
}
\value{
eset including only the supplied sample names.
}
\description{
Sets up esets for \link[MAMA]{MetaArray-class}. An eset is generated using
only the supplied sample names. Used to generate eset for each contrast
within a GSE.
}
\seealso{
\code{\link{clean_y}}.
}

