% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{get_raw}
\alias{get_raw}
\title{Download and unpack microarray supplementary files from GEO.}
\usage{
get_raw(gse_names, data_dir)
}
\arguments{
\item{gse_names}{Character vector of GSE names to download.}

\item{data_dir}{base data directory (a folder for each GSE will be created
here).}
}
\value{
NULL (for download/unpack only).
}
\description{
Downloads and unpacks microarray supplementary files from GEO.
Files are stored in the supplied data directory under the GSE name.
}
\examples{
\dontrun{

}

}

