% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{get_top_table}
\alias{get_top_table}
\title{Get top table}
\usage{
get_top_table(
  lm_fit,
  groups = c("test", "ctrl"),
  with.es = TRUE,
  robust = FALSE,
  trend = FALSE,
  allow.no.resid = FALSE
)
}
\arguments{
\item{lm_fit}{Result of \link{run_limma}}

\item{groups}{Test and Control group as strings.}

\item{with.es}{Add \code{'dprime'} and \code{'vardprime'} from 
\code{\link[metaMA]{effectsize}}? Default is \code{TRUE}.}

\item{robust}{logical, should the estimation of \code{df.prior} and \code{var.prior} be robustified against outlier sample variances?}

\item{trend}{logical, should an intensity-trend be allowed for the prior variance? Default is that the prior variance is constant.}

\item{allow.no.resid}{Allow no residual degrees of freedom? if \code{TRUE} and
the fit contrast matrix has no residual degrees of freedom, \link[limma]{eBayes} fit is skipped 
and the result of \link[limma]{contrasts.fit} is returned.}
}
\value{
result of \link[limma]{toptable}
}
\description{
Get top table
}
