% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vsd.R
\name{get_vsd}
\alias{get_vsd}
\title{Get variance stabilized data for exploratory data analysis}
\usage{
get_vsd(eset, rlog_cutoff = 50)
}
\arguments{
\item{eset}{ExpressionSet loaded with \link{load_raw}.
Requires group column in \code{pData(eset)} specifying sample groupings.}

\item{rlog_cutoff}{Sample number above which will use
\code{\link[DESeq2]{vst}} instead of \code{\link[DESeq2]{rlog}}.
Default is 50.}
}
\value{
\code{DESeqTransform} with variance stabilized expression data.
}
\description{
Get variance stabilized data for exploratory data analysis
}
