% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expr.R
\name{iqr_replicates}
\alias{iqr_replicates}
\title{Removes features with replicated annotation.}
\usage{
iqr_replicates(eset, annot = "SYMBOL", rm.dup = FALSE, keep_path = NULL)
}
\arguments{
\item{eset}{Annotated eset created by \code{load_raw}.}

\item{annot}{feature to use to remove replicates.}

\item{rm.dup}{remove duplicates (same measure, multiple ids)? Used for Pathway analysis so that doesn't treat
probes that map to multiple genes as distinct measures.}

\item{keep_path}{Path to file to load/save rows that are retained. Used for
caching by 'dseqr' app.}
}
\value{
Expression set with unique features at probe or gene level.
}
\description{
For rows with duplicated annot, highested IQR retained.
}
