% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_utils.R
\name{load_diff}
\alias{load_diff}
\title{Load previous differential expression analysis.}
\usage{
load_diff(gse_names, data_dir, probe = FALSE)
}
\arguments{
\item{gse_names}{Character vector specifying GSE names to be loaded.}

\item{data_dir}{String specifying directory of GSE folders.}

\item{probe}{Load probe level analysis? Default loads gene level analysis.}
}
\value{
Result of previous call to \code{diff_expr}.
}
\description{
Previous runs of \code{diff_expr} are loaded.
}
\examples{
library(lydata)
library(crossmeta)

data_dir <- system.file("extdata", package = "lydata")
gse_names<- c("GSE9601", "GSE34817")
prev <- load_diff(gse_names, data_dir)
}
\seealso{
\code{\link{diff_expr}}.
}

