% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_utils.R
\name{load_diff}
\alias{load_diff}
\title{Load previous differential expression analysis.}
\usage{
load_diff(gse_names, data_dir, probe = F)
}
\arguments{
\item{gse_names}{Character vector of GSE names to be loaded.}

\item{data_dir}{base data directory containing a folder for each GSE name.}

\item{probe}{Load probe level analysis? If FALSE, loads gene level analysis.}
}
\value{
saved results from previous call to \code{diff_expr}.
}
\description{
Previous runs of diff_expr are loaded.
}
\examples{

}
\seealso{
\link{diff_expr}
}

