% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{load_raw}
\alias{load_raw}
\title{Load and annotate raw data downloaded from GEO.}
\usage{
load_raw(gse_names, data_dir = getwd())
}
\arguments{
\item{gse_names}{Character vector of GSE names.}

\item{data_dir}{String specifying directory with GSE folders.}
}
\value{
List of annotated esets.
}
\description{
Loads and annotates raw data previously downloaded with \code{get_raw}.
Supported platforms include Affymetrix, Agilent, and Illumina.
}
\examples{
library(lydata)
data_dir <- system.file("extdata", package = "lydata")
eset <- load_raw("GSE9601", data_dir)
}
\seealso{
\code{\link{get_raw}} to obtain raw data.
}

