% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMA_utils.R
\name{make_ma}
\alias{make_ma}
\title{Make MetaArray object from results of differential expression.}
\usage{
make_ma(diff_exprs, sva)
}
\arguments{
\item{diff_exprs}{Result from \code{diff_expr} or \code{load_diff}.}

\item{sva}{Use esets with effect of surrogate variables removed? If FALSE,
non-adjusted esets will be used.}
}
\value{
MetaArray object.
}
\description{
Function is used to construct a MetaArray object from results of
\code{diff_expr}. MetaArray object allows for use of meta-analysis methods
present in MAMA package.
}
\seealso{
\link[MAMA]{MetaArray-class}, \link{diff_expr}, and \link{load_diff}.
}

