% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{match_prev_eset}
\alias{match_prev_eset}
\title{Reuse contrast selections from previous analysis.}
\usage{
match_prev_eset(eset, prev_anal)
}
\arguments{
\item{eset}{expression set passed to diff_expr.}

\item{prev_anal}{result of previous call to diff_expr.}
}
\value{
expression set with samples and pData as in prev_anal.
}
\description{
Transfers user-supplied selections from previous call of diff_expr.
}
\seealso{
\code{\link{diff_expr}}
}

