% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_illum.R
\name{open_raw_illum}
\alias{open_raw_illum}
\title{Open raw Illumina microarray files.}
\usage{
open_raw_illum(gse_names, data_dir = getwd())
}
\arguments{
\item{gse_names}{Character vector of Illumina GSE names to open.}

\item{data_dir}{String specifying directory with GSE folders.}
}
\value{
Character vector of successfully formated Illumina GSE names.
}
\description{
Helper function to open raw Illumina microarray files in order to check that
they are formatted correctly. For details on correct format, please see
'Checking Raw Illumina Data' in vignette.
}
\examples{
library(lydata)

# Illumina GSE names
illum_names <- c("GSE50841", "GSE34817", "GSE29689")

# location of raw data
data_dir <- system.file("extdata", package = "lydata")

# open raw data files with default text editor
# open_raw_illum(illum_names)
}
