% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_select_contrasts.R
\name{run_select_contrasts}
\alias{run_select_contrasts}
\title{Shiny gadget to upload groups and select contrasts}
\usage{
run_select_contrasts(
  eset,
  gse_name,
  prev = NULL,
  app_dir = system.file("select_contrasts", package = "crossmeta", mustWork = TRUE),
  port = 3838
)
}
\arguments{
\item{eset}{ExpressionSet}

\item{gse_name}{GEO accession for the series.}

\item{prev}{Previous result of \code{diff_expr}. Used to allow rechecking previous selections.}

\item{app_dir}{Directory to shiny app. For local development use 'inst/select_contrasts'. Default is
in 'select_contrasts' sub directory of crossmeta package.}

\item{port}{See \code{\link[shiny:shiny-package]{runApp()}}.}
}
\value{
result of \link{setup_prev}. Used to specify sample groups and contrasts for differential expression analysis.
}
\description{
Shiny gadget to upload groups and select contrasts
}
\keyword{"GSE9601"}
\keyword{"lydata")}
\keyword{#}
\keyword{<-}
\keyword{=}
\keyword{GSE}
\keyword{analysis}
\keyword{data}
\keyword{data_dir}
\keyword{data_dir)[[1]]}
\keyword{eset}
\keyword{gather}
\keyword{gse_name}
\keyword{gse_name)}
\keyword{internal}
\keyword{library(lydata)}
\keyword{load}
\keyword{load_raw(gse_name,}
\keyword{location}
\keyword{names}
\keyword{of}
\keyword{package}
\keyword{previous}
\keyword{run_select_contrasts(eset,}
\keyword{system.file("extdata",}
\keyword{z}
