% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_utils.R
\name{run_sva}
\alias{run_sva}
\title{Run surrogate variable analysis}
\usage{
run_sva(mods, eset, svanal)
}
\arguments{
\item{mods}{result of \code{get_mods}}

\item{eset}{ExpressionSet}

\item{svanal}{Should surrogate variable analysis be run? If \code{FALSE}, returns dummy result.}
}
\description{
Run surrogate variable analysis
}
