% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_utils.R
\name{symbol_annot}
\alias{symbol_annot}
\title{Adds gene name (SYMBOL) column to fData slot of expression set.}
\usage{
symbol_annot(eset, gpl_name)
}
\arguments{
\item{eset}{expression set object to add annotation data to.}

\item{gpl_name}{platform name used to look up bioconductor
annotation data package.}
}
\value{
eset with gene names (in "SYMBOL" column of fData slot).
}
\description{
Function uses platform (GPL) to identify and download corresponding
bioconductor annotation data package. Gene name ("SYMBOL") is then added to
featureData slot of supplied eset.
}
\examples{
\dontrun{

}
}
\seealso{
\link{load_affy}, \link{load_illum}, \link{load_agil}.
}

